+++ core/src/nucleus/ribosome/api/call.rs
@@ -4,7 +4,7 @@ use holochain_core_types::error::HolochainError;
 use holochain_dna::zome::capabilities::Membrane;
 use instance::RECV_DEFAULT_TIMEOUT_MS;
 use nucleus::{
-    get_capability_with_zome_call, launch_zome_fn_call, ribosome::api::Runtime,
+    get_capability_with_zome_call, launch_zome_fn_call, ribosome::runtime::Runtime,
     state::NucleusState, ZomeFnCall,
 };
 use serde_json;
