+++ core/src/nucleus/ribosome/api/get_entry.rs
@@ -1,6 +1,6 @@
 use futures::executor::block_on;
 use holochain_wasm_utils::api_serialization::get_entry::{GetEntryArgs, GetEntryResult};
-use nucleus::{actions::get_entry::get_entry, ribosome::api::Runtime};
+use nucleus::{actions::get_entry::get_entry, ribosome::runtime::Runtime};
 use serde_json;
 use wasmi::{RuntimeArgs, RuntimeValue, Trap};
 
@@ -54,8 +54,8 @@ mod tests {
     };
     use instance::tests::{test_context_and_logger, test_instance};
     use nucleus::{
+        ribosome,
         ribosome::api::{
-            call,
             commit::tests::test_commit_args_bytes,
             tests::{test_capability, test_parameters, test_zome_name},
         },
@@ -172,7 +172,7 @@ mod tests {
             "commit_dispatch",
             &test_parameters(),
         );
-        let call_result = call(
+        let call_result = ribosome::run_dna(
             &dna.name.to_string(),
             Arc::clone(&context),
             wasm.clone(),
@@ -194,7 +194,7 @@ mod tests {
             "get_dispatch",
             &test_parameters(),
         );
-        let call_result = call(
+        let call_result = ribosome::run_dna(
             &dna.name.to_string(),
             Arc::clone(&context),
             wasm.clone(),
@@ -238,7 +238,7 @@ mod tests {
             "get_dispatch",
             &test_parameters(),
         );
-        let call_result = call(
+        let call_result = ribosome::run_dna(
             &dna.name.to_string(),
             Arc::clone(&context),
             wasm.clone(),
